partial refactor to fix undefined import issue
