+++ src/test/web.test.js
@@ -18,9 +18,6 @@ import { groupTest, groupTestPartial } from './lib/test-lib'
 import * as regex from '../web'
 import { goodDomainNames, badDomainNames } from './data/domains'
 import { goodFQDomainNames, badFQDomainNames } from './data/fq-domains'
-import { goodIPs, badIPs } from './data/ips'
-import { goodIPFormats, badIPFormats } from './data/ip-formats'
-import { goodIPV6s, badIPV6s } from './data/ipv6s'
 import { goodTLDs, badTLDs } from './data/tlds'
 import { 
     goodFileUrls, badFileUrls,
@@ -34,15 +31,6 @@ groupTestPartial(regex.domainLabelREString, goodDomainNames, badDomainNames, 'su
 
 groupTest(regex.fqDomainNameRE, goodFQDomainNames, badFQDomainNames, 'FQ domain names')
 groupTestPartial(regex.fqDomainNameREString, goodFQDomainNames, badFQDomainNames, 'FQ domain names', undefined, undefined, 'u')
-
-groupTest(regex.ipHostRE, goodIPs, badIPs, 'IP address')
-groupTestPartial(regex.ipHostREString, goodIPs, badIPs, 'IP address')
-
-groupTest(regex.ipAddressRE, goodIPFormats, badIPFormats, 'IP format')
-groupTestPartial(regex.ipAddressREString, goodIPFormats, badIPFormats, 'IP format')
-
-groupTest(regex.ipV6RE, goodIPV6s, badIPV6s, 'IPV6')
-groupTestPartial(regex.ipV6REString, goodIPV6s, badIPV6s, 'IPV6')
 
 const goodLocalhosts = ['127.0.0.1', '127.34.0.255', 'localhost', '::1', '0:0:0:0:0:0:0:1']
 const badLocalhosts = ['128.0.0.1', 'localhosts', '::2']
