+++ src/web.js
@@ -15,54 +15,11 @@ limitations under the License.
 */
 
 import { emailREString } from './contacts'
+import { ipHostREString, ipV6REString, ipVFutureREString, ipTupleREString } from './ips'
 import { lockdownRE } from './lib/lockdown-re'
 import { uniNonASCII } from './lib/uni-non-ascii'
 
-// IP address dotted notation octets
-// excludes loopback network 0.0.0.0
-// excludes reserved space >= 224.0.0.0
-// excludes network & broacast addresses
-// (first & last IP address of each class)
-export const ipHostREString = '(?:[1-9]\\d?|1\\d\\d|2[01]\\d|22[0-3])' +
-  '(?:\\.(?:1?\\d{1,2}|2[0-4]\\d|25[0-5])){2}' +
-  '(?:\\.(?:[1-9]\\d?|1\\d\\d|2[0-4]\\d|25[0-4]))'
-// Web: Matches a valid, non-localhost IP address.
-export const ipHostRE = lockdownRE(ipHostREString)
-
-const ipTuple = '(?:0|1?\\d{1,2}|2[0-4]\\d|25[0-5])'
-export const ipAddressREString = `(?:${ipTuple}\\.){3}${ipTuple}`
-// Web: Matches a string in IP address format. Use 'ipHostRE' to match actually valid IP addresses.
-export const ipAddressRE = lockdownRE(ipAddressREString)
-
-// credit to: https://stackoverflow.com/a/17871737/929494
-export const ipV6REString =
-  '(([0-9a-fA-F]{1,4}:){7,7}[0-9a-fA-F]{1,4}|' + // 1:2:3:4:5:6:7:8
-  '([0-9a-fA-F]{1,4}:){1,7}:|' + // 1::                              1:2:3:4:5:6:7::
-  '([0-9a-fA-F]{1,4}:){1,6}:[0-9a-fA-F]{1,4}|' + // 1::8             1:2:3:4:5:6::8  1:2:3:4:5:6::8
-  '([0-9a-fA-F]{1,4}:){1,5}(:[0-9a-fA-F]{1,4}){1,2}|' + // 1::7:8           1:2:3:4:5::7:8  1:2:3:4:5::8
-  '([0-9a-fA-F]{1,4}:){1,4}(:[0-9a-fA-F]{1,4}){1,3}|' + // 1::6:7:8         1:2:3:4::6:7:8  1:2:3:4::8
-  '([0-9a-fA-F]{1,4}:){1,3}(:[0-9a-fA-F]{1,4}){1,4}|' + // 1::5:6:7:8       1:2:3::5:6:7:8  1:2:3::8
-  '([0-9a-fA-F]{1,4}:){1,2}(:[0-9a-fA-F]{1,4}){1,5}|' + // 1::4:5:6:7:8     1:2::4:5:6:7:8  1:2::8
-  '[0-9a-fA-F]{1,4}:((:[0-9a-fA-F]{1,4}){1,6})|' + // 1::3:4:5:6:7:8   1::3:4:5:6:7:8  1::8
-  ':((:[0-9a-fA-F]{1,4}){1,7}|:)|' + // ::2:3:4:5:6:7:8  ::2:3:4:5:6:7:8 ::8       ::
-  // (link-local IPv6 addresses with zone index)
-  'fe80:(:[0-9a-fA-F]{0,4}){0,4}%[0-9a-zA-Z]{1,}|' + // fe80::7:8%eth0   fe80::7:8%1
-  '::(ffff(:0{1,4}){0,1}:){0,1}' +
-  '((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\\.){3,3}' +
-  // (IPv4-mapped IPv6 addresses and IPv4-translated addresses)
-  '(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])|' + // ::255.255.255.255   ::ffff:255.255.255.255  ::ffff:0:255.255.255.255
-  '([0-9a-fA-F]{1,4}:){1,4}:' +
-  '((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\\.){3,3}' +
-  // 2001:db8:3:4::192.0.2.33  64:ff9b::192.0.2.33 (IPv4-Embedded IPv6 Address)
-  '(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9]))'
-// Web: Matches a string in IPV6 format.
-export const ipV6RE = lockdownRE(ipV6REString)
-
-export const ipVFutureREString = '(?:v[0-9a-fA-F]+\\.[a-zA-Z0-9~_.!$&\'()*+,;=:-]+)'
-// Web: Matches potential future IP protocols.
-export const ipVFutureRE = lockdownRE(ipVFutureREString)
-
-export const localhostREString = `(?:localhost|127(?:\\.${ipTuple}){3}|::1|0:0:0:0:0:0:0:1)`
+export const localhostREString = `(?:localhost|127(?:\\.${ipTupleREString}){3}|::1|0:0:0:0:0:0:0:1)`
 // Web: Matches any representation of localhost; the special name, IPV4 loopbacks, or IPV6 loopbacks.
 export const localhostRE = lockdownRE(localhostREString)
